% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acme_est.R
\name{acme.est}
\alias{acme.est}
\title{Parameter Estimation for Persistence and Search Proficiency}
\usage{
acme.est(rd, fname = "acme.est")
}
\arguments{
\item{rd}{list output from acme::read.data()}

\item{fname}{file name to which the output parameters are saved}
}
\value{
\code{acme.est} returns a list with the following components:
\item{params}{5-element vector: alpha and rho parameters for the
Weibull persistence distribution, a and b parameters for the
exponentially decreasing search proficiency, and bt as the bleed-through
rate}
\item{info}{list of select system information inherited from
acme::read.data() output}
}
\description{
Finds Maximum Likelihood Estimates Weibull persistence parameters,
and for exponentially decreasing search proficiency.
}

