% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rst.R
\name{Rst}
\alias{Rst}
\title{Evalute ACME Reduction Factor R*}
\usage{
Rst(Iij = 7, arabt = c(alp = 0.4695, rho = 0.0809, a = 1.0322, b = 0.0706,
  bt = 0.9573), kmax = 5, v = FALSE)
}
\arguments{
\item{Iij}{Search interval length (days)}

\item{arabt}{5-element vector: alpha and rho parameters for the Weibull
persistence distribution, a and b parameters for the Exponential search
proficiency distribution, and bt as the bleed-through parameter}

\item{kmax}{number of intervals to use in calculation - includes current
interval and number of look-back intervals. Minimum is 1 (only current
interval).}

\item{v}{logical. Verbose flag - see Value for what is reported}
}
\value{
If verbose, returns upper bound on truncation error (total and as a
fraction of R*), R* calculated from only current interval, and expected
 fraction of "old" carcasses discovered. If not verbose, returns only R*.
}
\description{
Calculates R*, the reduction factor (inverse of inflation factor for mortality
estimates), based on parameter estimates and maximum number of previous
search intervals to consider.
}

